refactor: resolve circular reference of Checkbox
