+++ components/checkbox/Group.tsx
@@ -4,6 +4,7 @@ import * as React from 'react';
 import { ConfigContext } from '../config-provider';
 import type { CheckboxChangeEvent } from './Checkbox';
 import Checkbox from './Checkbox';
+import GroupContext from './GroupContext';
 
 import useStyle from './style';
 
@@ -33,17 +34,6 @@ export interface CheckboxGroupProps extends AbstractCheckboxGroupProps {
   onChange?: (checkedValue: Array<CheckboxValueType>) => void;
   children?: React.ReactNode;
 }
-
-export interface CheckboxGroupContext {
-  name?: string;
-  toggleOption?: (option: CheckboxOptionType) => void;
-  value?: any;
-  disabled?: boolean;
-  registerValue: (val: string) => void;
-  cancelValue: (val: string) => void;
-}
-
-export const GroupContext = React.createContext<CheckboxGroupContext | null>(null);
 
 const InternalCheckboxGroup: React.ForwardRefRenderFunction<HTMLDivElement, CheckboxGroupProps> = (
   {
